requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.9.2
supabase==2.9.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
apscheduler==3.10.4
pytz==2023.3

# Para desarrollo y testing
pytest==7.4.3
httpx==0.25.2

# Para mejor performance
orjson==3.9.10

# Para logs y monitoreo
structlog==23.2.0

# Para validación adicional
email-validator==2.1.0

# Para mapas y cálculos geográficos
geopy==2.4.0

# Para cache y performance
cachetools==5.3.2

# Para mejores tipos de datos
typing-extensions==4.8.0 